Refactor NeoFaker.Person
