+++ lib/neo_faker/id_id/person/utils.ex
@@ -29,12 +29,7 @@ defmodule NeoFaker.IdId.Person.Utils do
 
     formatted_year = year |> to_string() |> String.slice(2, 2)
     formatted_month = month |> to_string() |> String.pad_leading(2, "0")
-
-    formatted_day =
-      [day, day + 40]
-      |> Enum.random()
-      |> to_string()
-      |> String.pad_leading(2, "0")
+    formatted_day = [day, day + 40] |> Enum.random() |> to_string() |> String.pad_leading(2, "0")
 
     "#{formatted_day}#{formatted_month}#{formatted_year}"
   end
